+++ src/lib/ParseApp.js
@@ -39,7 +39,8 @@ export default class ParseApp {
     serverInfo,
     production,
     iconName,
-    backgroundColor,
+    primaryBackgroundColor,
+    secondaryBackgroundColor,
     supportedPushLocales,
   }) {
     this.name = appName;
@@ -61,7 +62,8 @@ export default class ParseApp {
     this.serverURL = serverURL;
     this.serverInfo = serverInfo;
     this.icon = iconName;
-    this.backgroundColor=backgroundColor;
+    this.primaryBackgroundColor=primaryBackgroundColor;
+    this.secondaryBackgroundColor=secondaryBackgroundColor;
     this.supportedPushLocales = supportedPushLocales ? supportedPushLocales : [];
 
     if(!supportedPushLocales) {